requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
pydantic
python-dotenv
celery[redis]
redis
requests
